Fix: Reduce install size with --no-install-reccomends. Add: stop_server CLI command.
